2e Major opgelost. int16 veranderd in int32

+++ Breda Maps/Breda Maps/View/MainPage.xaml.cs
@@ -99,7 +99,7 @@ namespace Breda_Maps.View
                 double num3 = Math.Pow(Math.Sin(num2 / 2), 2) + Math.Cos(latitude) * Math.Cos(num) * Math.Pow(Math.Sin(num1 / 2), 2);
                 double num4 = 2 * Math.Atan2(Math.Sqrt(num3), Math.Sqrt(1 - num3));
                 double num5 = 6376500 * num4;
-                int intdist = Convert.ToInt16(num5);
+                int intdist = Convert.ToInt32(num5);
                 return (RoundNum(intdist));
             }
         }
